Separate contribution workflow information into two parts #120
The CONTRIBUTING.md describes the contribution workflow combining the
information for how to set up and manage a fork and for how to create
and process a pull request.

This change separates the two parts of the contribution process into
separate sections. In addition, is extends the explanation of how to
manage the local clone for a fork by an alternative workflow of having
the main Eclipse repository as the origin remote.

To conform to recent repository merges, the change also clean ups the
list of repositories in the Eclipse platform organization.

Contributes to https://github.com/eclipse-platform/.github/issues/120