Streamline CONTRIBUTING.md information on PR creation #120
The contributing information on how to create a PR is a rather verbose and slightly unstructred list of entires, which may not necessarily be relevant in the current order. This change adds a reference to the GitHub documentation on the standard "pull request from a fork" workflow that we follow. It also structures the workflow into five steps (preparing, creating, updating, validating and merging a PR), reducing the general information given by the standard workflow and putting more focus on Eclipse-platform-specific information.

Contributes to https://github.com/eclipse-platform/.github/issues/120